+++ src/exifmwg/ImageMetadata.hpp
@@ -7,10 +7,9 @@
 
 #include "DimensionsStruct.hpp"
 #include "KeywordInfoModel.hpp"
+#include "PythonBindable.hpp"
 #include "RegionInfoStruct.hpp"
 #include "XmpAreaStruct.hpp"
-
-namespace fs = std::filesystem;
 
 class ImageMetadata {
 public:
@@ -29,6 +28,12 @@ public:
   std::optional<std::string> City;
   std::optional<std::string> State;
   std::optional<std::string> Location;
+
+  ImageMetadata() = default;
+
+  ImageMetadata(const std::filesystem::path& path);
+
+  // This is used mostly in Python level testing, to construct expected structures
   ImageMetadata(int imageHeight, int imageWidth, std::optional<std::string> title = std::nullopt,
                 std::optional<std::string> description = std::nullopt,
                 std::optional<RegionInfoStruct> regionInfo = std::nullopt,
@@ -40,8 +45,28 @@ public:
                 std::optional<KeywordInfoModel> keywordInfo = std::nullopt,
                 std::optional<std::string> country = std::nullopt, std::optional<std::string> city = std::nullopt,
                 std::optional<std::string> state = std::nullopt, std::optional<std::string> location = std::nullopt);
+
+  void toFile(const std::optional<std::filesystem::path>& newPath = std::nullopt);
+  static void clearFile(const std::filesystem::path& path);
+
+  // Python bindable
+  std::string to_string() const;
+
+  friend bool operator==(const ImageMetadata& lhs, const ImageMetadata& rhs) {
+    return (lhs.ImageHeight == rhs.ImageHeight) && (lhs.ImageWidth == rhs.ImageWidth) && (lhs.Title == rhs.Title) &&
+           (lhs.Description == rhs.Description) && (lhs.RegionInfo == rhs.RegionInfo) &&
+           (lhs.Orientation == rhs.Orientation) && (lhs.LastKeywordXMP == rhs.LastKeywordXMP) &&
+           (lhs.TagsList == rhs.TagsList) && (lhs.CatalogSets == rhs.CatalogSets) &&
+           (lhs.HierarchicalSubject == rhs.HierarchicalSubject) && (lhs.KeywordInfo == rhs.KeywordInfo) &&
+           (lhs.Country == rhs.Country) && (lhs.City == rhs.City) && (lhs.State == rhs.State) &&
+           (lhs.Location == rhs.Location);
+  }
+
+private:
+  std::optional<std::filesystem::path> m_originalPath;
 };
 
 // Equality operators
-bool operator==(const ImageMetadata& lhs, const ImageMetadata& rhs);
-bool operator!=(const ImageMetadata& lhs, const ImageMetadata& rhs);
+static_assert(std::copy_constructible<ImageMetadata>);
+static_assert(std::equality_comparable<ImageMetadata>);
+static_assert(PythonBindableRepr<ImageMetadata>);