+++ src/exifmwg/KeywordInfoModel.hpp
@@ -5,6 +5,7 @@
 
 #include <exiv2/exiv2.hpp>
 
+#include "PythonBindable.hpp"
 #include "XmpSerializable.hpp"
 
 class KeywordInfoModel {
@@ -23,6 +24,13 @@ public:
     static KeywordStruct fromXmp(const Exiv2::XmpData& xmpData, const std::string& basePath);
     void toXmp(Exiv2::XmpData& xmpData, const std::string& basePath) const;
 
+    // Python bindable
+    std::string to_string() const;
+
+    friend bool operator==(const KeywordStruct& lhs, const KeywordStruct& rhs) {
+      return (lhs.Keyword == rhs.Keyword) && (lhs.Applied == rhs.Applied) && (lhs.Children == rhs.Children);
+    }
+
   private:
     void writeChildrenToXmp(Exiv2::XmpData& xmpData, const std::string& basePath) const;
   };
@@ -37,9 +45,16 @@ public:
   static KeywordInfoModel fromXmp(const Exiv2::XmpData& xmpData);
   void toXmp(Exiv2::XmpData& xmpData) const;
 
+  // Python bindable
+  std::string to_string() const;
+
   // Operators
   KeywordInfoModel& operator|=(const KeywordInfoModel& other);
   KeywordInfoModel operator|(const KeywordInfoModel& other) const;
+
+  friend bool operator==(const KeywordInfoModel& lhs, const KeywordInfoModel& rhs) {
+    return lhs.Hierarchy == rhs.Hierarchy;
+  }
 
 private:
   std::vector<KeywordStruct> mergeKeywordVectors(const std::vector<KeywordStruct>& vec1,
@@ -48,8 +63,12 @@ private:
   std::optional<bool> mergeApplied(const std::optional<bool>& a, const std::optional<bool>& b);
 };
 
-bool operator==(const KeywordInfoModel::KeywordStruct& lhs, const KeywordInfoModel::KeywordStruct& rhs);
-bool operator==(const KeywordInfoModel& lhs, const KeywordInfoModel& rhs);
-
-static_assert(XmpSerializable<KeywordInfoModel>);
+static_assert(std::copy_constructible<KeywordInfoModel::KeywordStruct>);
+static_assert(std::equality_comparable<KeywordInfoModel::KeywordStruct>);
 static_assert(XmpSerializableWithKey<KeywordInfoModel::KeywordStruct>);
+static_assert(PythonBindableRepr<KeywordInfoModel::KeywordStruct>);
+
+static_assert(std::copy_constructible<KeywordInfoModel>);
+static_assert(std::equality_comparable<KeywordInfoModel>);
+static_assert(XmpSerializable<KeywordInfoModel>);
+static_assert(PythonBindableRepr<KeywordInfoModel>);