+++ src/exifmwg/Logging.cpp
@@ -1,4 +1,4 @@
-#include "utils.hpp"
+#include "Logging.hpp"
 
 #include <iostream>