+++ src/exifmwg/__init__.py
@@ -1,3 +1,5 @@
+from exifmwg.bindings import EXIV2_VERSION
+from exifmwg.bindings import EXPAT_VERSION
 from exifmwg.bindings import Dimensions
 from exifmwg.bindings import ImageMetadata
 from exifmwg.bindings import Keyword
@@ -5,13 +7,10 @@ from exifmwg.bindings import KeywordInfo
 from exifmwg.bindings import Region
 from exifmwg.bindings import RegionInfo
 from exifmwg.bindings import XmpArea
-from exifmwg.bindings import clear_existing_metadata
-from exifmwg.bindings import exiv2_version
-from exifmwg.bindings import expat_version
-from exifmwg.bindings import read_metadata
-from exifmwg.bindings import write_metadata
 
 __all__ = [
+    "EXIV2_VERSION",
+    "EXPAT_VERSION",
     "Dimensions",
     "ImageMetadata",
     "Keyword",
@@ -19,9 +18,4 @@ __all__ = [
     "Region",
     "RegionInfo",
     "XmpArea",
-    "clear_existing_metadata",
-    "exiv2_version",
-    "expat_version",
-    "read_metadata",
-    "write_metadata",
 ]