+++ src/exifmwg/reading.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <filesystem>
-
-#include "ImageMetadata.hpp"
-
-ImageMetadata read_metadata(const std::filesystem::path& filepath);